finish thread binding
