+++ python/tvm/script/tir/scope_handler.py
@@ -16,7 +16,7 @@
 # under the License.
 """TVM Script Parser Scope Handler Classes"""
 # pylint: disable=redefined-builtin, unused-argument, invalid-name, relative-beyond-top-level
-from typing import Tuple, Any, Callable, Optional, List, Union, Mapping
+from typing import Tuple, Any, Callable, Optional, List, Union, Mapping, overload
 
 import synr
 import tvm.tir
@@ -500,9 +500,12 @@ class Serial(ForScopeHandler):
     def __init__(self):
         def serial(
             begin: PrimExpr,
-            end: PrimExpr,
+            end: PrimExpr = None,
             annotations: Optional[Mapping[str, Object]] = None,
         ):
+            if not end:
+                end = begin
+                begin = 0
             self.create_loop_info(begin, end, ForKind.SERIAL, annotations=annotations)
 
         super().__init__(serial)
@@ -515,9 +518,12 @@ class Parallel(ForScopeHandler):
     def __init__(self):
         def parallel(
             begin: PrimExpr,
-            end: PrimExpr,
+            end: PrimExpr = None,
             annotations: Optional[Mapping[str, Object]] = None,
         ):
+            if not end:
+                end = begin
+                begin = 0
             self.create_loop_info(begin, end, ForKind.PARALLEL, annotations=annotations)
 
         super().__init__(parallel)
@@ -530,9 +536,12 @@ class Vectorized(ForScopeHandler):
     def __init__(self):
         def vectorized(
             begin: PrimExpr,
-            end: PrimExpr,
+            end: PrimExpr = None,
             annotations: Optional[Mapping[str, Object]] = None,
         ):
+            if not end:
+                end = begin
+                begin = 0
             self.create_loop_info(begin, end, ForKind.VECTORIZED, annotations=annotations)
 
         super().__init__(vectorized)
@@ -545,9 +554,12 @@ class Unroll(ForScopeHandler):
     def __init__(self):
         def unroll(
             begin: PrimExpr,
-            end: PrimExpr,
+            end: PrimExpr = None,
             annotations: Optional[Mapping[str, Object]] = None,
         ):
+            if not end:
+                end = begin
+                begin = 0
             self.create_loop_info(begin, end, ForKind.UNROLLED, annotations=annotations)
 
         super().__init__(unroll)
@@ -560,10 +572,19 @@ class ThreadBinding(ForScopeHandler):
     def __init__(self):
         def thread_binding(
             begin: PrimExpr,
-            end: PrimExpr,
-            thread: str,
+            end: PrimExpr = None,
+            thread: str = None,
             annotations: Optional[Mapping[str, Object]] = None,
         ):
+            if not thread:
+                if isinstance(end, str):  # handle case like thread_binding(128, "threadIdx.x")
+                    thread = end
+                    end = None
+                else:
+                    raise ValueError("Thread cannot be None for thread_binding")
+            if not end:
+                end = begin
+                begin = 0
             thread_iter_var = IterVar(None, None, IterVar.ThreadIndex, thread)
             self.create_loop_info(
                 begin,
